requirements_ws.txt
attrdict==2.0.0
gevent==1.4.0
gevent-websocket==0.10.1
h5py==2.6.0
python-socketio==4.4.0
numpy==1.16.0
Pillow==2.2.1
scipy==0.19.1
keras==2.0.8
tensorflow-gpu==1.15.0
ujson==1.35
wsaccel==0.6.2